requirements.txt
Django==5.2.1
djangorestframework==3.16.0
pandas==2.2.3